image_captioning_segmentation_project_yolov8/requirements.txt
streamlit
pillow
numpy
opencv-python
torch
torchvision
transformers

torch
torchvision

ultralytics
